+++ OMOMoney/Data/SwiftData/OMOMoneySchema.swift
@@ -3,7 +3,7 @@ import SwiftData
 
 enum SchemaV1: VersionedSchema {
     static var versionIdentifier = Schema.Version(1, 0, 0)
-    
+
     static var models: [any PersistentModel.Type] {
         [
             SDUser.self,